requirements.txt
gunicorn
dash==1.11.0
pandas==1.1.4
mysql.connector